infra/requirements.txt
aws-cdk-lib==2.15.0
constructs==10.1.6
